test: split UAT regression tests into per-doctype files with a shared base case
