+++ buildsuite_core/buildsuite_core/doctype/work_package/test_work_package.py
@@ -1,8 +1,10 @@
 # Copyright (c) 2026, Infraholic Innovations Pvt. Ltd and Contributors
 # See license.txt
 
-# import frappe
+import frappe
 from frappe.tests import IntegrationTestCase
+
+from buildsuite_core.tests.base import BuildSuiteTestCase
 
 
 # On IntegrationTestCase, the doctype test records and all
@@ -12,11 +14,47 @@ EXTRA_TEST_RECORD_DEPENDENCIES = []  # eg. ["User"]
 IGNORE_TEST_RECORD_DEPENDENCIES = []  # eg. ["User"]
 
 
-
 class IntegrationTestWorkPackage(IntegrationTestCase):
-	"""
-	Integration tests for WorkPackage.
-	Use this class for testing interactions between multiple components.
-	"""
-
 	pass
+
+
+class TestWorkPackage(BuildSuiteTestCase):
+	"""Work Package status auto-advances from its tasks' progress, but a manual
+	On Hold is respected."""
+
+	def test_wp_status_autoadvances(self):
+		p = self._make_project(company=self.company)
+		wp = frappe.get_doc({
+			"doctype": "Work Package", "project": p.name,
+			"work_package_name": "UAT WP", "code": f"WP-{self._n}", "status": "Planned",
+		}).insert(ignore_permissions=True)
+		t1 = frappe.get_doc({
+			"doctype": "Task", "subject": "WP task 1", "project": p.name,
+			"work_package": wp.name, "task_status": "Yet To Start",
+		}).insert(ignore_permissions=True)
+		t2 = frappe.get_doc({
+			"doctype": "Task", "subject": "WP task 2", "project": p.name,
+			"work_package": wp.name, "task_status": "Yet To Start",
+		}).insert(ignore_permissions=True)
+
+		# Any task progress → Planned advances to In Progress.
+		self._file_tpe(t1.name, 40)
+		self.assertEqual(frappe.db.get_value("Work Package", wp.name, "status"), "In Progress")
+
+		# All tasks at 100 → Completed.
+		self._file_tpe(t1.name, 100)
+		self._file_tpe(t2.name, 100)
+		self.assertEqual(frappe.db.get_value("Work Package", wp.name, "status"), "Completed")
+
+	def test_wp_status_on_hold_is_respected(self):
+		p = self._make_project(company=self.company)
+		wp = frappe.get_doc({
+			"doctype": "Work Package", "project": p.name,
+			"work_package_name": "UAT WP hold", "code": f"WPH-{self._n}", "status": "On Hold",
+		}).insert(ignore_permissions=True)
+		t = frappe.get_doc({
+			"doctype": "Task", "subject": "held task", "project": p.name,
+			"work_package": wp.name, "task_status": "Yet To Start",
+		}).insert(ignore_permissions=True)
+		self._file_tpe(t.name, 40)  # progress filed, but WP is manually On Hold
+		self.assertEqual(frappe.db.get_value("Work Package", wp.name, "status"), "On Hold")
