--- buildsuite_core/tests/test_uat_fixes.py
@@ -1,529 +1,0 @@
-# Copyright (c) 2026, Infraholic Innovations Pvt. Ltd and contributors
-# For license information, please see license.txt
-"""Regression tests for the M1 UAT failure fixes.
-
-Each test maps to one or more UAT cases (referenced in the test name/docstring).
-UnitTestCase wraps every test in a transaction and rolls it back, so the records
-created here never persist.
-"""
-
-import frappe
-from frappe.tests import UnitTestCase
-from frappe.utils import today
-
-
-def _company():
-	return frappe.db.get_value("Company", {}, "name")
-
-
-class TestUATFixes(UnitTestCase):
-	def setUp(self):
-		self.company = _company()
-		self._n = frappe.generate_hash(length=6)
-
-	# --- helpers ---------------------------------------------------------
-	def _make_project(self, status="Ongoing", parent=None, company=None):
-		doc = frappe.get_doc({
-			"doctype": "Project",
-			"project_name": f"UAT {self._n}",
-			"custom_project_id": f"UAT-{self._n}",
-			"project_status": status,
-			"parent_project": parent,
-			"company": company,
-		})
-		doc.insert(ignore_permissions=True)
-		return doc
-
-	def _make_task(self, project, task_status="Yet To Start"):
-		doc = frappe.get_doc({
-			"doctype": "Task",
-			"subject": f"UAT Task {frappe.generate_hash(length=4)}",
-			"project": project,
-			"task_status": task_status,
-		})
-		doc.insert(ignore_permissions=True)
-		return doc
-
-	def _file_tpe(self, task, pct, blocker=0, blocker_detail=None):
-		doc = frappe.get_doc({
-			"doctype": "Task Progress Entry",
-			"task": task,
-			"entry_date": today(),
-			"cumulative_progress": pct,
-			"blocker": blocker,
-			"blocker_detail": blocker_detail,
-		})
-		doc.insert(ignore_permissions=True)
-		return doc
-
-	# --- C: project status sync (PRJ-008, PRJ-009) ----------------------
-	def test_project_status_syncs_to_native(self):
-		p = self._make_project(status="Ongoing", company=self.company)
-		self.assertEqual(p.project_status, "Ongoing")
-		self.assertEqual(p.status, "Open")
-
-		p.project_status = "Completed"
-		p.save(ignore_permissions=True)
-		self.assertEqual(p.status, "Completed")
-
-	def test_project_status_defaults_to_new(self):
-		# PRJ-008: blank project_status becomes New on save.
-		p = frappe.get_doc({
-			"doctype": "Project",
-			"project_name": f"UAT {self._n}",
-			"custom_project_id": f"UAT-{self._n}",
-			"company": self.company,
-		})
-		p.insert(ignore_permissions=True)
-		self.assertEqual(p.project_status, "New")
-		self.assertEqual(p.status, "Open")
-
-	# --- E: company inherit + lock (PRJ-005, PRJ-013) -------------------
-	def test_subproject_inherits_company(self):
-		parent = self._make_project(company=self.company)
-		child = frappe.get_doc({
-			"doctype": "Project",
-			"project_name": f"UAT sub {self._n}",
-			"custom_project_id": f"UAT-SUB-{self._n}",
-			"parent_project": parent.name,
-			"is_group": 0,
-		})
-		child.insert(ignore_permissions=True)
-		self.assertEqual(child.company, self.company)
-
-	def test_project_infers_company_from_creating_user(self):
-		# PRJ-012/013: company isn't sent by the form; it's inferred from the
-		# creating user's company.
-		frappe.db.set_value("User", frappe.session.user, "company", self.company)
-		p = frappe.get_doc({
-			"doctype": "Project",
-			"project_name": f"UAT infer {self._n}",
-			"custom_project_id": f"UAT-INF-{self._n}",
-		})
-		p.insert(ignore_permissions=True)
-		self.assertEqual(p.company, self.company)
-
-	def test_user_persona_company_optional(self):
-		# Company is optional for persona'd users (the earlier mandatory-company
-		# rule was dropped). A persona'd user with no company must still save.
-		u = frappe.new_doc("User")
-		u.email = f"uat-{self._n}@example.com"
-		u.first_name = "UAT"
-		u.user_type = "System User"
-		u.send_welcome_email = 0
-		u.persona = "Project Manager"
-		u.company = ""
-		u.insert(ignore_permissions=True)  # must not raise
-		self.assertEqual(u.persona, "Project Manager")
-
-	def test_user_persona_with_company_ok(self):
-		u = frappe.get_doc({
-			"doctype": "User",
-			"email": f"uat-ok-{self._n}@example.com",
-			"first_name": "UAT",
-			"user_type": "System User",
-			"send_welcome_email": 0,
-			"persona": "Project Manager",
-			"company": self.company,
-		})
-		u.insert(ignore_permissions=True)
-		self.assertEqual(u.company, self.company)
-
-	def test_company_locked_after_create(self):
-		# PRJ-013: company can't be changed on edit; revert is silent.
-		others = frappe.get_all("Company", filters={"name": ("!=", self.company)}, pluck="name")
-		p = self._make_project(company=self.company)
-		if others:
-			p.company = others[0]
-			p.save(ignore_permissions=True)
-			self.assertEqual(p.company, self.company)
-
-	# --- B: task status sync (TSK-006, TSK-007) -------------------------
-	def test_task_status_syncs_to_native(self):
-		p = self._make_project(company=self.company)
-		t = self._make_task(p.name, task_status="Yet To Start")
-		self.assertEqual(t.status, "Open")
-
-		t.task_status = "In Progress"
-		t.save(ignore_permissions=True)
-		self.assertEqual(t.status, "Working")
-
-	# --- A + F: TPE updates task + project, no invalid status throw ------
-	def test_tpe_updates_task_status_and_no_throw(self):
-		# TPE-006: 40% -> In Progress (previously threw on project.status="Working").
-		p = self._make_project(company=self.company)
-		t = self._make_task(p.name)
-		self._file_tpe(t.name, 40)
-		t.reload()
-		self.assertEqual(t.progress, 40)
-		self.assertEqual(t.task_status, "In Progress")
-
-	def test_tpe_100_completes_task(self):
-		# TPE-007
-		p = self._make_project(company=self.company)
-		t = self._make_task(p.name)
-		self._file_tpe(t.name, 100)
-		t.reload()
-		self.assertEqual(t.progress, 100)
-		self.assertEqual(t.task_status, "Completed")
-
-	# --- F: project progress rollup (PRJ-010) ---------------------------
-	def test_project_progress_rolls_up(self):
-		p = self._make_project(company=self.company)
-		t1 = self._make_task(p.name)
-		t2 = self._make_task(p.name)
-		self._file_tpe(t1.name, 100)
-		self._file_tpe(t2.name, 50)
-		p.reload()
-		self.assertEqual(p.percent_complete, 75)
-
-	def test_project_progress_rounded_to_whole_number(self):
-		# 3 tasks at 33/33/34 average to 33.33… — the rollup must store a whole number.
-		p = self._make_project(company=self.company)
-		for pct in (33, 33, 34):
-			self._file_tpe(self._make_task(p.name).name, pct)
-		p.reload()
-		self.assertEqual(p.percent_complete, round(p.percent_complete))
-		self.assertEqual(p.percent_complete, 33)
-
-	def test_parent_progress_weighted_by_subproject_task_count(self):
-		# Parent progress blends its direct tasks (weight 1 each) with each
-		# subproject's progress weighted by the subproject's task count.
-		parent = self._make_project(company=self.company)
-		parent.is_group = 1
-		parent.save(ignore_permissions=True)
-
-		# Parent direct task at 60%.
-		pt = self._make_task(parent.name)
-		self._file_tpe(pt.name, 60)
-
-		# Subproject with 3 tasks all at 100% -> sub progress 100, weight 3.
-		sub = frappe.get_doc({
-			"doctype": "Project",
-			"project_name": f"UAT sub {self._n}",
-			"custom_project_id": f"UAT-SUB-{self._n}",
-			"parent_project": parent.name,
-			"is_group": 0,
-		}).insert(ignore_permissions=True)
-		for _ in range(3):
-			st = self._make_task(sub.name)
-			self._file_tpe(st.name, 100)
-
-		sub.reload()
-		self.assertEqual(sub.percent_complete, 100)
-
-		# (60*1 + 100*3) / (1 + 3) = 90
-		parent.reload()
-		self.assertEqual(parent.percent_complete, 90)
-
-	# --- G: TPE delete reverts task (TPE-009, TPE-010) ------------------
-	def test_delete_latest_tpe_reverts_to_previous(self):
-		p = self._make_project(company=self.company)
-		t = self._make_task(p.name)
-		self._file_tpe(t.name, 40)
-		latest = self._file_tpe(t.name, 100)
-		t.reload()
-		self.assertEqual(t.progress, 100)
-
-		frappe.delete_doc("Task Progress Entry", latest.name, ignore_permissions=True)
-		t.reload()
-		self.assertEqual(t.progress, 40)
-		self.assertEqual(t.task_status, "In Progress")
-
-	def test_delete_only_tpe_reverts_to_zero(self):
-		p = self._make_project(company=self.company)
-		t = self._make_task(p.name)
-		only = self._file_tpe(t.name, 60)
-		frappe.delete_doc("Task Progress Entry", only.name, ignore_permissions=True)
-		t.reload()
-		self.assertEqual(t.progress, 0)
-		self.assertEqual(t.task_status, "Yet To Start")
-
-	# --- J: blocker requires a note (TPE-011) ---------------------------
-	def test_blocker_requires_note(self):
-		p = self._make_project(company=self.company)
-		t = self._make_task(p.name)
-		with self.assertRaises(frappe.ValidationError):
-			self._file_tpe(t.name, 30, blocker=1, blocker_detail=None)
-
-	def test_blocker_with_note_ok(self):
-		p = self._make_project(company=self.company)
-		t = self._make_task(p.name)
-		tpe = self._file_tpe(t.name, 30, blocker=1, blocker_detail="Rain stopped the pour")
-		self.assertTrue(tpe.name)
-
-	# --- TPE progress is monotonic — can't go below current ----------------
-	def test_tpe_rejects_below_current_progress(self):
-		p = self._make_project(company=self.company)
-		t = self._make_task(p.name)
-		self._file_tpe(t.name, 50)
-		before = frappe.db.count("Task Progress Entry", {"task": t.name})
-
-		with self.assertRaises(frappe.ValidationError):
-			self._file_tpe(t.name, 40)  # below current 50% → rejected
-
-		# Rejected entry must NOT have been logged.
-		self.assertEqual(frappe.db.count("Task Progress Entry", {"task": t.name}), before)
-		t.reload()
-		self.assertEqual(t.progress, 50)
-
-	def test_tpe_equal_and_increase_allowed(self):
-		p = self._make_project(company=self.company)
-		t = self._make_task(p.name)
-		self._file_tpe(t.name, 50)
-		self._file_tpe(t.name, 50)   # equal — allowed
-		self._file_tpe(t.name, 70)   # increase — allowed
-		t.reload()
-		self.assertEqual(t.progress, 70)
-
-	# --- Work Package status auto-advances from task progress --------------
-	def test_wp_status_autoadvances(self):
-		p = self._make_project(company=self.company)
-		wp = frappe.get_doc({
-			"doctype": "Work Package", "project": p.name,
-			"work_package_name": "UAT WP", "code": f"WP-{self._n}", "status": "Planned",
-		}).insert(ignore_permissions=True)
-		t1 = frappe.get_doc({
-			"doctype": "Task", "subject": "WP task 1", "project": p.name,
-			"work_package": wp.name, "task_status": "Yet To Start",
-		}).insert(ignore_permissions=True)
-		t2 = frappe.get_doc({
-			"doctype": "Task", "subject": "WP task 2", "project": p.name,
-			"work_package": wp.name, "task_status": "Yet To Start",
-		}).insert(ignore_permissions=True)
-
-		# Any task progress → Planned advances to In Progress.
-		self._file_tpe(t1.name, 40)
-		self.assertEqual(frappe.db.get_value("Work Package", wp.name, "status"), "In Progress")
-
-		# All tasks at 100 → Completed.
-		self._file_tpe(t1.name, 100)
-		self._file_tpe(t2.name, 100)
-		self.assertEqual(frappe.db.get_value("Work Package", wp.name, "status"), "Completed")
-
-	def test_wp_status_on_hold_is_respected(self):
-		p = self._make_project(company=self.company)
-		wp = frappe.get_doc({
-			"doctype": "Work Package", "project": p.name,
-			"work_package_name": "UAT WP hold", "code": f"WPH-{self._n}", "status": "On Hold",
-		}).insert(ignore_permissions=True)
-		t = frappe.get_doc({
-			"doctype": "Task", "subject": "held task", "project": p.name,
-			"work_package": wp.name, "task_status": "Yet To Start",
-		}).insert(ignore_permissions=True)
-		self._file_tpe(t.name, 40)  # progress filed, but WP is manually On Hold
-		self.assertEqual(frappe.db.get_value("Work Package", wp.name, "status"), "On Hold")
-
-	# --- Project team add/remove (custom_team_members) ------------------
-	def test_project_team_add_and_remove(self):
-		from buildsuite_core.api.project_team import (
-			add_project_team_member,
-			remove_project_team_member,
-		)
-		p = self._make_project(company=self.company)
-
-		team = add_project_team_member(p.name, "Administrator")
-		self.assertTrue(any(m["user"] == "Administrator" for m in team))
-
-		# Idempotent — adding again doesn't duplicate.
-		team = add_project_team_member(p.name, "Administrator")
-		self.assertEqual(len([m for m in team if m["user"] == "Administrator"]), 1)
-
-		team = remove_project_team_member(p.name, "Administrator")
-		self.assertFalse(any(m["user"] == "Administrator" for m in team))
-
-	# --- I: guarded cascade delete (PRJ-014, TSK-013) -------------------
-	def test_cascade_delete_removes_descendants(self):
-		p = self._make_project(company=self.company)
-		t = self._make_task(p.name)
-		tpe = self._file_tpe(t.name, 20)
-		wp = frappe.get_doc({
-			"doctype": "Work Package", "project": p.name,
-			"work_package_name": "UAT WP", "code": f"WP-{self._n}",
-		}).insert(ignore_permissions=True)
-
-		frappe.delete_doc("Project", p.name, ignore_permissions=True)
-
-		self.assertFalse(frappe.db.exists("Project", p.name))
-		self.assertFalse(frappe.db.exists("Task", t.name))
-		self.assertFalse(frappe.db.exists("Task Progress Entry", tpe.name))
-		self.assertFalse(frappe.db.exists("Work Package", wp.name))
-
-	# --- Task assignee = single-assignee _assign -------------------------
-	def test_task_assignee_is_single_via_assign(self):
-		"""Assignee maps to Frappe-native _assign with single-assignee semantics:
-		reassigning drops the previous user; clearing empties it."""
-		from buildsuite_core.api.task_assignment import set_task_assignee, get_task_assignee
-
-		p = self._make_project(company=self.company)
-		t = self._make_task(p.name)
-		users = frappe.get_all(
-			"User", filters={"enabled": 1, "user_type": "System User"},
-			pluck="name", limit=2,
-		)
-		u1, u2 = users[0], users[1]
-
-		set_task_assignee(t.name, u1)
-		self.assertEqual(get_task_assignee(t.name), u1)
-
-		# Reassigning leaves exactly one open assignment, pointing at the new user.
-		set_task_assignee(t.name, u2)
-		self.assertEqual(get_task_assignee(t.name), u2)
-		open_todos = frappe.get_all(
-			"ToDo",
-			filters={"reference_type": "Task", "reference_name": t.name, "status": ("!=", "Cancelled")},
-			pluck="allocated_to",
-		)
-		self.assertEqual(open_todos, [u2])
-
-		set_task_assignee(t.name, None)
-		self.assertIsNone(get_task_assignee(t.name))
-
-	# --- Site Engineer submits own Stage Planning for approval -----------
-	def test_site_engineer_submits_own_stage_planning(self):
-		"""A Site Engineer must be able to apply the Submit transition to their
-		own Draft stage. apply_workflow sets workflow_state to the next state
-		before saving, so the own-scope write gate must check the PERSISTED
-		state, not the in-memory one — otherwise the submit self-locks."""
-		from frappe.model.workflow import apply_workflow
-
-		se = f"se-{self._n}@example.com"
-		frappe.get_doc({
-			"doctype": "User", "email": se, "first_name": "SE",
-			"send_welcome_email": 0, "user_type": "System User",
-			"persona": "Site Engineer", "company": self.company,
-		}).insert(ignore_permissions=True)
-
-		p = frappe.get_doc({
-			"doctype": "Project", "project_name": f"WF {self._n}",
-			"custom_project_id": f"WF-{self._n}", "project_status": "Ongoing",
-			"company": self.company, "custom_team_members": [{"user": se}],
-		}).insert(ignore_permissions=True)
-
-		frappe.set_user(se)
-		try:
-			st = frappe.get_doc({
-				"doctype": "Stage Planning", "stage_name": f"WF {self._n}",
-				"project": p.name, "workflow_state": "Draft",
-			}).insert()
-			apply_workflow(st, "Submit for Approval")
-			st.reload()
-			self.assertEqual(st.workflow_state, "Pending Approval")
-
-			# Once submitted, the SE can no longer edit their own (now locked) stage.
-			st.stage_name = "should not save"
-			self.assertRaises(frappe.PermissionError, st.save)
-		finally:
-			frappe.set_user("Administrator")
-
-	# --- Stage approval UAT (SAW-006 / SAW-011 / SAW-013) ----------------
-	def _make_stage(self, project, task=None, state="Draft"):
-		rows = [{"task": task, "planned_qty": 80, "qty_unit": "%"}] if task else []
-		return frappe.get_doc({
-			"doctype": "Stage Planning",
-			"stage_name": f"ST {frappe.generate_hash(length=4)}",
-			"project": project,
-			"workflow_state": state,
-			"stage_planning_tasks": rows,
-		}).insert(ignore_permissions=True)
-
-	def _approve(self, st):
-		from frappe.model.workflow import apply_workflow
-		apply_workflow(st, "Submit for Approval")
-		apply_workflow(st, "Approve")
-		st.reload()
-
-	def test_revise_clones_rejected_stage(self):
-		"""SAW-006 — revising a Rejected stage clones it into a fresh Draft and
-		leaves the original Rejected stage untouched (audit record)."""
-		from frappe.model.workflow import apply_workflow
-		from buildsuite_core.buildsuite_core.doctype.stage_planning.stage_planning import (
-			reject_stage_planning, revise_stage_planning,
-		)
-		p = self._make_project()
-		t = self._make_task(p.name)
-		st = self._make_stage(p.name, task=t.name)
-		apply_workflow(st, "Submit for Approval")
-		reject_stage_planning(st.name, "Scope unclear")
-		st.reload()
-		self.assertEqual(st.workflow_state, "Rejected")
-
-		res = revise_stage_planning(st.name)
-		clone = frappe.get_doc("Stage Planning", res["name"])
-		self.assertNotEqual(clone.name, st.name)
-		self.assertEqual(clone.workflow_state, "Draft")
-		self.assertFalse(clone.reject_reason)
-		self.assertEqual(len(clone.stage_planning_tasks), 1)
-		self.assertEqual(clone.stage_planning_tasks[0].task, t.name)
-		# Original stays Rejected.
-		st.reload()
-		self.assertEqual(st.workflow_state, "Rejected")
-
-	def test_approved_stage_task_edit_throws(self):
-		"""SAW-011 — an Approved stage's task list / planned qty is locked."""
-		p = self._make_project()
-		t = self._make_task(p.name)
-		st = self._make_stage(p.name, task=t.name)
-		self._approve(st)
-		st.stage_planning_tasks[0].planned_qty = 50
-		self.assertRaises(frappe.ValidationError, lambda: st.save(ignore_permissions=True))
-
-	def test_approved_to_draft_allows_task_change(self):
-		"""SAW-011 — Revise (Approved -> Draft) reopens editing."""
-		from frappe.model.workflow import apply_workflow
-		p = self._make_project()
-		t = self._make_task(p.name)
-		st = self._make_stage(p.name, task=t.name)
-		self._approve(st)
-		apply_workflow(st, "Revise")
-		st.reload()
-		self.assertEqual(st.workflow_state, "Draft")
-		st.stage_planning_tasks[0].planned_qty = 50
-		st.save(ignore_permissions=True)  # must not raise
-		st.reload()
-		self.assertEqual(st.stage_planning_tasks[0].planned_qty, 50)
-
-	def test_stage_transition_logs_activity(self):
-		"""SAW-013 — each workflow transition is recorded on the timeline."""
-		from frappe.model.workflow import apply_workflow
-		from buildsuite_core.buildsuite_core.doctype.stage_planning.stage_planning import (
-			get_stage_activity,
-		)
-		p = self._make_project()
-		st = self._make_stage(p.name)
-		apply_workflow(st, "Submit for Approval")
-		types = [a["type"] for a in get_stage_activity(st.name)]
-		self.assertIn("created", types)
-		self.assertIn("submitted", types)
-
-	# --- PRM-002: PM auto-added to team ---------------------------------
-	def test_project_manager_auto_added_to_team(self):
-		pm = f"pm-{self._n}@example.com"
-		frappe.get_doc({
-			"doctype": "User", "email": pm, "first_name": "PM",
-			"send_welcome_email": 0, "user_type": "System User",
-			"persona": "Project Manager", "company": self.company,
-		}).insert(ignore_permissions=True)
-		p = frappe.get_doc({
-			"doctype": "Project", "project_name": f"PMTEAM {self._n}",
-			"custom_project_id": f"PMTEAM-{self._n}", "project_status": "Ongoing",
-			"company": self.company, "project_manager": pm,
-		}).insert(ignore_permissions=True)
-		self.assertTrue(frappe.db.exists("Project Team", {
-			"parent": p.name, "parentfield": "custom_team_members", "user": pm,
-		}))
-
-	# --- PTT-005: seed stages onto an existing project ------------------
-	def test_seed_stages_from_template(self):
-		if not frappe.db.exists("BuildSuite Project Template", {"project_type": "Commercial"}):
-			self.skipTest("No Commercial template seeded on this site")
-		from buildsuite_core.utils.project import seed_stages_from_template
-		p = self._make_project()
-		frappe.db.set_value("Project", p.name, "project_type", "Commercial")
-		res = seed_stages_from_template(p.name)
-		self.assertGreater(res["seeded"], 0)
-		stages = frappe.get_all("Stage Planning", filters={"project": p.name}, pluck="name")
-		self.assertTrue(stages)
-		for s in stages:
-			for r in frappe.get_doc("Stage Planning", s).stage_planning_tasks:
-				self.assertEqual(r.planned_qty, 100)
